# Flags: remove protocol sniffing flags