--- a/manifests/charts/istiod-remote/values.yaml
+++ b/manifests/charts/istiod-remote/values.yaml
@@ -31,10 +31,6 @@ pilot:
   volumeMounts: []
   # Additional volumes to the istiod pod
   volumes: []
-  # if protocol sniffing is enabled for outbound
-  enableProtocolSniffingForOutbound: true
-  # if protocol sniffing is enabled for inbound
-  enableProtocolSniffingForInbound: true
   nodeSelector: {}
   podAnnotations: {}
   serviceAnnotations: {}