--- a/pilot/pkg/networking/networking_test.go
+++ b/pilot/pkg/networking/networking_test.go
@@ -19,123 +19,63 @@ import (
 
 	core "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
 
-	"istio.io/istio/pilot/pkg/features"
 	"istio.io/istio/pkg/config/protocol"
-	"istio.io/istio/pkg/test"
 )
 
 func TestModelProtocolToListenerProtocol(t *testing.T) {
 	tests := []struct {
-		name                       string
-		protocol                   protocol.Instance
-		direction                  core.TrafficDirection
-		sniffingEnabledForInbound  bool
-		sniffingEnabledForOutbound bool
-		want                       ListenerProtocol
+		name      string
+		protocol  protocol.Instance
+		direction core.TrafficDirection
+		want      ListenerProtocol
 	}{
 		{
 			"TCP to TCP",
 			protocol.TCP,
 			core.TrafficDirection_INBOUND,
-			true,
-			true,
 			ListenerProtocolTCP,
 		},
 		{
 			"HTTP to HTTP",
 			protocol.HTTP,
 			core.TrafficDirection_INBOUND,
-			true,
-			true,
 			ListenerProtocolHTTP,
 		},
 		{
 			"HTTP to HTTP",
 			protocol.HTTP_PROXY,
 			core.TrafficDirection_OUTBOUND,
-			true,
-			true,
 			ListenerProtocolHTTP,
 		},
 		{
 			"MySQL to TCP",
 			protocol.MySQL,
 			core.TrafficDirection_INBOUND,
-			true,
-			true,
 			ListenerProtocolTCP,
 		},
 		{
 			"Inbound unknown to Auto",
 			protocol.Unsupported,
 			core.TrafficDirection_INBOUND,
-			true,
-			true,
 			ListenerProtocolAuto,
 		},
 		{
 			"Outbound unknown to Auto",
 			protocol.Unsupported,
 			core.TrafficDirection_OUTBOUND,
-			true,
-			true,
-			ListenerProtocolAuto,
-		},
-		{
-			"Inbound unknown to TCP",
-			protocol.Unsupported,
-			core.TrafficDirection_INBOUND,
-			false,
-			true,
-			ListenerProtocolTCP,
-		},
-		{
-			"Outbound unknown to Auto (disable sniffing for inbound)",
-			protocol.Unsupported,
-			core.TrafficDirection_OUTBOUND,
-			false,
-			true,
-			ListenerProtocolAuto,
-		},
-		{
-			"Outbound unknown to Auto (disable sniffing for outbound)",
-			protocol.Unsupported,
-			core.TrafficDirection_OUTBOUND,
-			true,
-			false,
-			ListenerProtocolTCP,
-		},
-		{
-			"Inbound unknown to Auto (disable sniffing for outbound)",
-			protocol.Unsupported,
-			core.TrafficDirection_INBOUND,
-			true,
-			false,
 			ListenerProtocolAuto,
 		},
 		{
 			"UDP to UDP",
 			protocol.UDP,
 			core.TrafficDirection_INBOUND,
-			true,
-			false,
 			ListenerProtocolUnknown,
-		},
-		{
-			"Unknown Protocol",
-			"Bad Protocol",
-			core.TrafficDirection_INBOUND,
-			true,
-			false,
-			ListenerProtocolAuto,
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			test.SetForTest(t, &features.EnableProtocolSniffingForOutbound, tt.sniffingEnabledForOutbound)
-			test.SetForTest(t, &features.EnableProtocolSniffingForInbound, tt.sniffingEnabledForInbound)
-			if got := ModelProtocolToListenerProtocol(tt.protocol, tt.direction); got != tt.want {
+			if got := ModelProtocolToListenerProtocol(tt.protocol); got != tt.want {
 				t.Errorf("ModelProtocolToListenerProtocol() = %v, want %v", got, tt.want)
 			}
 		})