… manifests/profiles/default.yaml
… operator/pkg/apis/istio/v1alpha1/values_types.pb.go
… pilot/pkg/networking/core/v1alpha3/listener_test.go
… releasenotes/notes/flagprotocol-sniffing.yaml
… samples/addons/grafana.yaml