This has been enabled by default since 1.5 and is now a part of core Istio. Time to clean up all the dead code we have to support this flag. There will be more followup cleanup that this _enables_ but is not really realted to protocol sniffing so I left it ouut

TODO:
* Look for other places
* Add release notes
* Fix tests, probably