@istio-testing 2023-07-28
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
unit-tests-arm64_istio | ac82352c71c53aa1f18f4092153a00c1b7172076 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/46217/unit-tests-arm64_istio/1684719304961953792) | true | `/test unit-tests-arm64`
unit-tests_istio | ac82352c71c53aa1f18f4092153a00c1b7172076 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/46217/unit-tests_istio/1684719304915816448) | true | `/test unit-tests`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->