@howardjohn 2023-07-28
@hzxuzhonghu fyi, this will enable another PR  I am working on:

```
name                       old time/op        new time/op        delta
ListenerGeneration/http-8         123µs ±23%          99µs ±13%  -19.74%  (p=0.000 n=10+24)
ListenerGeneration/tcp-8          537µs ±24%         372µs ± 4%  -30.79%  (p=0.000 n=10+23)
ListenerGeneration/tls-8          550µs ±12%         386µs ± 5%  -29.77%  (p=0.000 n=10+25)
ListenerGeneration/auto-8        1.96ms ±13%        1.73ms ± 5%  -11.88%  (p=0.002 n=10+20)

name                       old alloc/op       new alloc/op       delta
ListenerGeneration/http-8         109kB ± 0%          74kB ± 0%  -32.20%  (p=0.000 n=10+25)
ListenerGeneration/tcp-8          424kB ± 0%         363kB ± 0%  -14.43%  (p=0.000 n=10+24)
ListenerGeneration/tls-8          425kB ± 0%         364kB ± 0%  -14.42%  (p=0.000 n=10+24)
ListenerGeneration/auto-8        1.44MB ± 0%        1.34MB ± 0%   -7.15%  (p=0.000 n=8+24)

name                       old allocs/op      new allocs/op      delta
ListenerGeneration/http-8         1.20k ± 0%         0.88k ± 0%  -26.11%  (p=0.000 n=10+25)
ListenerGeneration/tcp-8          5.98k ± 0%         4.38k ± 0%  -26.82%  (p=0.000 n=9+24)
ListenerGeneration/tls-8          5.99k ± 0%         4.38k ± 0%  -26.80%  (p=0.000 n=10+25)
ListenerGeneration/auto-8         16.7k ± 0%         14.7k ± 0%  -12.02%  (p=0.000 n=8+25)
```

Benchmark is not from this change (this one should have no impact), but it is enabled by the simplification here.